Use one SSE client; remove local repo poll re #12
