+++ client/src/index.js
@@ -89,6 +89,14 @@ function RouterElement({enableNet, setEnableNet, enabledRef}) {
     ]);
 
     const {enqueueSnackbar} = useSnackbar();
+    const netHandler = ev => {
+        if (ev.data === "enabled" && !enabledRef.current) {
+            setEnableNet(true);
+        } else if (ev.data === "disabled" && enabledRef.current) {
+            setEnableNet(false);
+        }
+    }
+
     const miscHandler = ev => {
         const dataBits = ev.data.split('--');
         if (dataBits.length === 4) {
@@ -103,35 +111,41 @@ function RouterElement({enableNet, setEnableNet, enabledRef}) {
     }
 
     useEffect(() => {
+        const controller = new AbortController();
         const fetchData = async () => {
-            await fetchEventSource("/notifications/misc", {
+            await fetchEventSource("/notifications", {
                 method: "GET",
                 headers: {
                     Accept: "text/event-stream",
                 },
                 onopen(res) {
                     if (res.ok && res.status === 200) {
-                        //console.log("Connected to misc SSE");
+                        //console.log("Connected to SSE");
                     } else if (
                         res.status >= 400 &&
                         res.status < 500 &&
                         res.status !== 429
                     ) {
-                        console.log("Error from misc SSE", res.status);
+                        console.log("Error from SSE", res.status);
                     }
                 },
                 onmessage(event) {
-                    miscHandler(event)
+                    if (event.event === "misc") {
+                        miscHandler(event)
+                    } else if (event.event === "net_status") {
+                        netHandler(event)
+                    }
                 },
                 onclose() {
-                    console.log("Misc SSE connection closed by the server");
+                    console.log("SSE connection closed by the server");
                 },
                 onerror(err) {
-                    console.log("There was an error from the Misc SSE server", err);
+                    console.log("There was an error from the SSE server", err);
                 },
             });
         };
         fetchData();
+        return () => controller.abort();
     }, []);
 
     return <Box sx={{height: '95vh', overflow: 'hidden'}}>
